Remove python client: lm_client/kvstore_client/fib_client (17/N)
Summary:
Finally the "RED" diff is coming for Python client removal in Open/R :P

After openr_audit script change to remove the usage of python client inside Open/R, we have reached the stage the NO python client usage for:

LinkMonitor, KvStore and Fib

This diff is to deprecate the usage of those python client and OpenrCtrlDeprecated class. We still yet to remove the opening of corresponding port in Open/R's binary since we still have customer using it.

Reviewed By: saifhhasan

Differential Revision: D16653471

fbshipit-source-id: a586ff47d50583dab201e50c4730d0f2c18441d7