Remove -Werror to NOT treat warning as error
Summary:
During the efforts to remove all of the .patch files in TG environment, we are facing some issues when compiling with the latest Open/R binary. It turns out to be that: CMAKE will treat warnings as ERRORs and stop the compilation.

As per discussion with John Strizich offline, should keep fbcode build and CMAKE build consistent when treating warnings.

Remove `-Werror` in CMAKE flags

Reviewed By: jstrizich

Differential Revision: D17172622

fbshipit-source-id: 09cd209465cbe7d066361b26219592feb7261346