Duplicate openr/nl library into fbmeshd/rnl
Summary:
Just copy them over and update the `#include`s such that they use each other. Most of the more complicated moving over happens in the diff above this.

We're moving out of `openr`, but have a significant dependency on this library. It's not great to copy it over, but since we are on a tight timeline to get out of `openr` this is the best thing we can do short term.

Reviewed By: ilyacodesFB

Differential Revision: D17295798

fbshipit-source-id: 907f68eb0b31f2bb6b60e271b90bd0d6eab94dd8