enbale KSPF algorithm for BGP routes
Summary:
Before, if we see BGP prefix with KSPF algorithm specified, we will see it as invalid route and ignore it.

In this diff, we enable KSPF algorithm with KSPF support because in turbo fabric POD aggregate will be learned as bgp routes.

I created a knob to control this feature to make sure it did not break current SPR setting.

Reviewed By: saifhhasan

Differential Revision: D17638183

fbshipit-source-id: 154c4e8726d5d6952fb917f0af365345fbc643b6