bgp emulation: linux_agent should return runstate of CONFIGURED
Summary:
Currently, linux_agent always returns SwitchRunState of UNINITIALIZED.  This
is a problem for BGP, because BGP only syncs routes when agent state >= CONFIGURED.
This issue is not seen if BGP is restarted when agent is already up, which
may be a different bug. T56585708 has been opened to track that issue.

Reviewed By: saifhhasan

Differential Revision: D18222186

fbshipit-source-id: 7a2a52052c342a1108b586a851a5a2692bf78e3b