Fix longPoll API with lspAgent integration test
Summary:
After integrating test with lspAgent from Ebb. We have observed longPoll API returns too frequently. Check it out and noticed that it is due to we check keysTobeUpdated field, which is NOT necessary. Check `reply->keyVals` should be good enough.

Checked with Saif Hasan's original comments and see this should be the right condition since lspAgent should always honor what Open/R has in KvStore.

P.S. also change the promise setValue(false) when it times out. This will NOT rely on client to catch the exception based on offline discussion with Rafit Izhak-Ratzin

Reviewed By: saifhhasan

Differential Revision: D18218243

fbshipit-source-id: 56cb54cc8b0d2c5691214f2143851d9d7034421a