Bug fix in ttl update for persistKey
Summary:
We had a bug where changing ttl for a key, already present in KvStore leads to
no-effect. Reason => Because we do not use new ttl value at all. This diff fixes
this bug.

Also, advertise ttl update immediately if ttl changes. This way even if value
doesn't change, we tell everyone in network to use new ttl value.

We found this bug during spr onboarding process, and ttl was changed to 1hr

NOTE: We can do better for persistKey, by re-advertising key which is expired
into the KvStore. Created T57424923 to track this feature.

Differential Revision: D18484889

fbshipit-source-id: c3a400682907724b956017e7422f96f9cce62b66