make backwards compatibilty for prefixDB more robust
Summary:
This code was a bit buggy. There are a number of edge cases that would result in us not having the proper prefix state.

Mainly what we will do on the decision side is always respect the KvStore state. e.g. if a full prefix db and per prefix db exist in the KvStore, we will merge the two. It is the originating node's responsibility to purge any stale state from the KvStore.

Differential Revision: D18816910

fbshipit-source-id: d96c31b6018049e1b1911661c3a41b348e10f5b4