fix bug in deleting expired keys
Summary: With per prefix keys enabled, we have a bug where we don't deleted the prefixes for a node that goes away. This doesn't cause bad routes but raises an alarm that we have no routes to these prefixes. The issue is with the behavior of `folly::split`. if there is a mismatch between the number of delineated fields and number of string arguments, the call is a no op.

Differential Revision: D19146760

fbshipit-source-id: c2c9064aff23aa1d3cc258ded5f662571a0add71