Wait for fibers to finish before stop
Summary: Introduce `addFiberTask` API in OpenrEventBase which will start a fiber and wait for it's future in stop sequence. This API ensures that all fibers MUST stop before an object is destroyed. This avoid spurious errors where fiber callback is executed in destruction sequence of EventBase and accessing freed memory locations.

Reviewed By: GirasoleY

Differential Revision: D19870708

fbshipit-source-id: 979a467c43d44efa0e798c0c113a5c9c1b5fbc18