fix flaky openr/link-monitor/tests:link_monitor_test - LinkMonitorTestFixture.DampenLinkFlaps
Summary:
the root cause is:
we initiate the linkMonitor and let it ran.
Then we stopped previous one and reinitiate a new linkMonitor.
The new linkmonitor is using same address for fbzmq. Sometimes, the resources is still held by the old LinkMonitor and causing the failure.

The fix is to change the API to let us instantiate a LinkMonitor with the require parameter, so that we don't have to re-create a new one.

Reviewed By: yi-xian

Differential Revision: D19910577

fbshipit-source-id: 33f11d9a85f67b52f017d6d13a1c950ebbe3b385