OPENR: reduce netlink interface timeout to 1s
Summary:
During fboss canary testing, when RSW was updated with new bundle images of
wedge_agent, bgp, and openr, we saw bgpmon issues on FSW, where it was
withdrawing the BGP routes originated by the RSW.

On looking at the logs, we see that when OPENR restarted, the initial call
to netlink from LinkMonitor failed, after a timeout of 20 seconds.
LinkMonitor then updated KvStore with 0 entries which results in the BGP
routes being temporarily withdrawn.  Here is a timeline:

09:00:46: OPENR on RSW007 gets SIGTERM and terminates itself
09:00:53: OPENR on RSW007 restarts
09:01:02: PrefixManager loads loopback addresses, 10.254.113.7/32 and fc00:0:0:1::6/128 from disk and populates KvStore
09:01:02: PrefixManager learns bgp routes from ::1
09:01:02: LinkMonitor thread starts and initiates interface sync from netlink
09:01:22: Netlink command times out (20 seconds!) and updates KvStore with empty interface list
09:01:22: PrefixManager withdraws loopback addresses
09:01:22: This results in BGP routes being withdrawn
09:01:25: LinkMonitor retries sync interface db
09:01:32: this retry took another 7 seconds (a total of 30 seconds from first try) but it succeeds

This leaves a window when the BGP routes get withdrawn because the loopback
address is deleted.  During this time, if the KvStore is queried by the FSW,
it will result in the BGP routes being witdrawn.  This is not seen every time.

Reviewed By: saifhhasan

Differential Revision: D20938059

fbshipit-source-id: aef784cd45933886fdc9a0cab681aad427494d97